requirements.txt
asgiref==3.9.1
Django==5.2.5
django-csp==4.0
django-environ==0.12.0
django-sslserver==0.22
dotenv==0.9.9
packaging==25.0
python-dotenv==1.1.1
sqlparse==0.5.3
